requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
APScheduler==3.10.4
openai>=1.40.0
google-generativeai==0.8.3
